Handle special case when 0 jet constituents by setting dummy objects
This is usually when we have a jet that gets promoted to top 2
via large JER. These events will normally fail selection
so it's OK later on, but as a placeholder we make dummy objects,
and return a dummy value.

Note that in the unfolding hists, we do a check for -ve values,
which will throw an error there.